// src/types/banner.ts
export type BannerType = {
    main: {
        id: string;
        title: string;
        description: string;
        image: string;
        link: string;
    }[];
    sub: {
        id: string;
        title: string;
        description: string;
        image: string;
        link: string;
    };
};